cursor类型由int改为long型

+++ EnigmaNet.DouYinOpenApi/IApiClient.cs
@@ -16,9 +16,9 @@ namespace EnigmaNet.DouYinOpenApi
 
         Task<Models.UserInfo> GetUserInfoAsync(string openId, string accessToken);
 
-        Task<Models.Following.FollowingListResult> GetFollowingListAsync(string openId, string accessToken, int pageSize, int? cursor);
+        Task<Models.Following.FollowingListResult> GetFollowingListAsync(string openId, string accessToken, int pageSize, long? cursor);
 
-        Task<Models.Fans.FansListResult> GetFansListAsync(string openId, string accessToken, int pageSize, int? cursor);
+        Task<Models.Fans.FansListResult> GetFansListAsync(string openId, string accessToken, int pageSize, long? cursor);
 
         Task<Models.Fans.DataResult> GetFansDataAsync(string openId, string accessToken);
 
@@ -26,15 +26,15 @@ namespace EnigmaNet.DouYinOpenApi
 
         Task<Models.Video.CreateVideoResult> CreateVideoAsync(string openId, string accessToken, string videoId, string text, string microAppId, string microAppTitle, string microAppUrl, double coverTime, string[] atUserOpenIds);
 
-        Task<Models.Video.VideoListResult> GetVideoListAsync(string openId, string accessToken, int pageSize, int? cursor);
+        Task<Models.Video.VideoListResult> GetVideoListAsync(string openId, string accessToken, int pageSize, long? cursor);
 
         Task<List<Models.Video.VideoItemInfo>> GetVideoDataAsync(string openId, string accessToken, string[] itemIds);
 
         Task DeleteVideoAsync(string openId, string accessToken, string[] itemIds);
 
-        Task<Models.Video.CommentListResult> GetVideoCommentListAsync(string openId, string accessToken, string itemId, int pageSize, int? cursor);
+        Task<Models.Video.CommentListResult> GetVideoCommentListAsync(string openId, string accessToken, string itemId, int pageSize, long? cursor);
 
-        Task<Models.Video.CommentListResult> GetVideoCommentReplyListAsync(string openId, string accessToken, string itemId, string commentId, int pageSize, int? cursor);
+        Task<Models.Video.CommentListResult> GetVideoCommentReplyListAsync(string openId, string accessToken, string itemId, string commentId, int pageSize, long? cursor);
 
         Task ReplyVideoCommentAsync(string openId, string accessToken, string itemId, string commentId, string content);
