+++ EnigmaNet.DouYinOpenApi/Models/Video/CommentListResult.cs
@@ -6,7 +6,7 @@ namespace EnigmaNet.DouYinOpenApi.Models.Video
 {
     public class CommentListResult
     {
-        public int Cursor { get; set; }
+        public long Cursor { get; set; }
         public bool HasMore { get; set; }
         public List<CommentInfo> List { get; set; }
     }
